Implement a stub API Client for the sake of getting info from cluster
A yaml parsed kube config is no longer needed, the kube_client crate
contains one that does everything we need. Still keeping the cli and
context specifyer rather than loaded the kubeconfig one because it would
be nice if this can be overridden from commandlines/show up in help etc.